partial add of cyber slayer materials
i compiled nearly all of this months ago when i originally uploaded the rest of the projects, but i was missing a couple things. forgot about it, and deleted my saved files for uploading here. looked over everything again realizing i forgot to upload everything so i gotta do it all again. this is pt 1